unify route schema
